requirements.txt
playwright>=1.41.0
PyYAML>=6.0.1
